nerine_server/requirements.txt
Django==1.11.2
djangorestframework==3.6.3
<<<<<<< HEAD
djangorestframework-jwk==1.11.0
=======
djangorestframework-jwt==1.11.0
PyJWT==1.5.2
>>>>>>> d43379b329012fb820dac3a27079a96f58f5b778
PyMySQL==0.7.11
pytz==2017.2
